Consider arch property of driver definitions
The bits property can no longer be used to distinguish between architectures, the new arch property is expected to be used for that use case. This property is optional and for many driver entries is not provided for the amd64/x86 binaries. Possibly need to only allow drivers with null arch for x86_64/amd64 and x86.